pyproject.toml
[tool.poetry]
name = "backend-tutorial"
version = "0.1.0"
description = ""
authors = ["SIGIRACE <chickang9@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "backend_tutorial"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
mysqlclient = "^2.1.1"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.168.1"}
strawberry-graphql-django = "^0.9.3"
pyjwt = "^2.6.0"
django-environ = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
